implement rabbitmq messaging

+++ MA.SlotService.Api/Contracts/SpinResultResponse.cs
@@ -4,7 +4,7 @@ public class SpinResultResponse
 {
     public required Guid SpinId { get; set; }
     
-    public required byte[] Result { get; set; }
+    public required int[] Result { get; set; }
     
     public required long Balance { get; init; }
 }
