+++ MA.SlotService.Application/Features/StartSpin/SpinResult.cs
@@ -4,5 +4,5 @@ public record SpinResult
 {
     public required Guid SpinId { get; set; }
     
-    public required byte[] Result { get; set; }
+    public required int[] Result { get; set; }
 }
